requirements.txt
pandas
scikit-learn
pyyaml
streamlit
